+++ build.rs
@@ -8,7 +8,6 @@ fn main() {
     let source_path = Path::new("test-framework").join("dist").join("framework.min.js");
     let dest_path = Path::new(&out);
     let dest_file_name = Path::new(&dest_path).join("framework.min.js");
-
     // println!("Copy framework file {} to {}", source_path.to_str().unwrap(), dest_file_name.to_str().unwrap());
     if !Path::exists(dest_path) {
         create_dir(dest_path).unwrap();